make core.config self-contained
